+++ examples/standing.php
@@ -2,8 +2,8 @@
 
 namespace Application;
 
-use MagicLegacy\Component\MtgMelee\Importer\Service\Client;
-use MagicLegacy\Component\MtgMelee\Importer\Service\Standing;
+use MagicLegacy\Component\MtgMelee\Client\MtgMeleeClient;
+use MagicLegacy\Component\MtgMelee\Client\Service\Standing;
 use Eureka\Component\Curl;
 use Nyholm\Psr7\Factory\Psr17Factory;
 
@@ -11,7 +11,7 @@ require_once __DIR__ . '/../vendor/autoload.php';
 
 //~ Declare tier required services (included as dependencies)
 $httpFactory    = new Psr17Factory();
-$mtgMeleeClient = new Client(
+$mtgMeleeClient = new MtgMeleeClient(
     new Curl\HttpClient(),
     $httpFactory,
     $httpFactory,