+++ CHANGELOG.md
@@ -10,3 +10,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
  * Add code to import and convert json tournament pairing data from MtgMelee into Entities
  * Add tests
  * Add doc
+
+
+## [0.2.0] 2020-08-28 - Release v0.2.0
+
+### Changed
+ * Rename `Client` to `MtgMeleeClient` and move it to root `src/`
+ * Rename composer name `magiclegacy/mtgmelee-importer` to `magiclegacy/mtgmelee-client`
+ * Update namespace of all class according to new naming
+ * Update example & README